# docs/openapi-order-service.yaml
openapi: 3.1.0
info:
  title: Order Service API
  version: 1.0.0
  description: API для управления заказами в системе печатного производства

servers:
  - url: http://localhost:8081
    description: Локальный сервер order-service

paths:
  /api/v1/orders:
    get:
      summary: Получить список заказов с фильтрами и пагинацией
      tags:
        - Order
      parameters:
        - name: status
          in: query
          description: Статус заказа
          schema:
            type: string
            enum: [WAITING, LAUNCHED, IN_PROGRESS, READY, ACCEPTED, CLOSED]
        - name: managerId
          in: query
          description: ID менеджера
          schema:
            type: integer
        - name: fromDate
          in: query
          description: Дата с (формат YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          description: Дата по (формат YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список заказов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPageResponse'

    post:
      summary: Создать новый заказ
      tags:
        - Order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Заказ создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'

  /api/v1/orders/{id}:
    get:
      summary: Получить детальную информацию о заказе
      tags:
        - Order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Детали заказа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailDto'
        '404':
          description: Заказ не найден

    put:
      summary: Обновить заказ (частичное обновление)
      tags:
        - Order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdateRequest'
      responses:
        '200':
          description: Заказ обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'

    delete:
      summary: Удалить заказ (soft delete)
      tags:
        - Order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Заказ удален

  /api/v1/orders/{id}/items:
    post:
      summary: Добавить позицию в заказ
      tags:
        - Order Items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItemCreateRequest'
      responses:
        '200':
          description: Позиция добавлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'

  /api/v1/orders/{id}/stages:
    post:
      summary: Добавить/загрузить этапы производства из шаблона
      tags:
        - Production
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StageTemplateRequest'
      responses:
        '200':
          description: Этапы добавлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'

  /api/v1/orders/{id}/payments:
    post:
      summary: Добавить оплату к заказу
      tags:
        - Finance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreateRequest'
      responses:
        '201':
          description: Оплата добавлена'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentDto'

components:
  schemas:
    OrderDetailDto:
      allOf:
        - $ref: '#/components/schemas/OrderDto'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/OrderItemDto'
            stages:
              type: array
              items:
                $ref: '#/components/schemas/OrderStageDto'
            payments:
              type: array
              items:
                $ref: '#/components/schemas/PaymentDto'
            comments:
              type: array
              items:
                $ref: '#/components/schemas/CommentDto'

    OrderDto:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        client:
          $ref: '#/components/schemas/ClientDto'
        description:
          type: string
        totalAmount:
          type: number
          format: decimal
        paidAmount:
          type: number
          format: decimal
        debtAmount:
          type: number
          format: decimal
        status:
          type: string
          enum: [WAITING, LAUNCHED, IN_PROGRESS, READY, ACCEPTED, CLOSED]
        productionStage:
          type: string
          enum: [NOT_STARTED, DESIGN, PRINTING, FINISHING, QUALITY_CONTROL, PACKAGING, SHIPPING]
        orderDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        manager:
          $ref: '#/components/schemas/EmployeeDto'
        hasDocuments:
          type: boolean
        createdAt:
          type: string
          format: date-time

    OrderCreateRequest:
      type: object
      required:
        - clientId
      properties:
        orderNumber:
          type: string
        clientId:
          type: integer
        description:
          type: string
        orderDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        managerId:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemCreateRequest'

    OrderItemCreateRequest:
      type: object
      required:
        - name
        - price
        - quantity
      properties:
        name:
          type: string
        price:
          type: number
          format: decimal
        quantity:
          type: integer
        readyDate:
          type: string
          format: date

    OrderItemDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
          format: decimal
        quantity:
          type: integer
        cost:
          type: number
          format: decimal
        readyDate:
          type: string
          format: date

    ClientDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [PRIVATE, COMPANY]
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string

    EmployeeDto:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        position:
          type: string
        phone:
          type: string
        email:
          type: string

    OrderStageDto:
      type: object
      properties:
        id:
          type: integer
        workshop:
          $ref: '#/components/schemas/WorkshopDto'
        waitPrevious:
          type: boolean
        dueDate:
          type: string
          format: date
        note:
          type: string
        status:
          type: string
          enum: [TODO, IN_PROGRESS, DONE]

    WorkshopDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sortOrder:
          type: integer

    PaymentDto:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
          format: decimal
        paymentDate:
          type: string
          format: date
        paymentType:
          type: string
        details:
          type: string
        isPartial:
          type: boolean

    PaymentCreateRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          format: decimal
        paymentDate:
          type: string
          format: date
        paymentType:
          type: string
        details:
          type: string
        isPartial:
          type: boolean

    CommentDto:
      type: object
      properties:
        id:
          type: integer
        message:
          type: string
        author:
          $ref: '#/components/schemas/EmployeeDto'
        isInternal:
          type: boolean
        timestamp:
          type: string
          format: date-time

    OrderPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/OrderDto'
        pageable:
          type: object
        totalElements:
          type: integer
        totalPages:
          type: integer
        last:
          type: boolean
        first:
          type: boolean
        size:
          type: integer
        number:
          type: integer
        sort:
          type: object
        numberOfElements:
          type: integer
        empty:
          type: boolean